requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
rich>=13.7.0
schedule>=1.2.0
openai>=1.0.0
python-dotenv>=1.0.0
jinja2>=3.1.0
flask>=3.0.0
selenium>=4.0.0
webdriver-manager>=4.0.0
PyPDF2>=3.0.0
python-docx>=1.0.0
